feat: bugs fixed, added remaining requirements

+++ apps/server/src/main/java/com/aayush/lad/hrms/modules/engagement/services/PostService.java
@@ -3,12 +3,13 @@ package com.aayush.lad.hrms.modules.engagement.services;
 import java.util.List;
 import java.util.UUID;
 
+import com.aayush.lad.hrms.core.exeptions.AccessDeniedException;
+import com.aayush.lad.hrms.modules.user.services.NotificationService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.aayush.lad.hrms.core.exeptions.ConflictException;
 import com.aayush.lad.hrms.core.exeptions.NotFoundException;
-import com.aayush.lad.hrms.core.exeptions.UnauthorisedException;
 import com.aayush.lad.hrms.core.services.CurrentUserService;
 import com.aayush.lad.hrms.modules.engagement.dtos.read.PostResponse;
 import com.aayush.lad.hrms.modules.engagement.dtos.write.CreateCommentRequest;
@@ -31,6 +32,7 @@ public class PostService {
     private final PostRepository postRepository;
     private final PostCommentRepository postCommentRepository;
     private final CurrentUserService currentUserService;
+    private final NotificationService notificationService;
     private final PostMapper mapper;
 
     public List<PostResponse> getAll() {
@@ -52,6 +54,11 @@ public class PostService {
     @Transactional
     public void update(UpdatePostRequest request) {
         Post post = getPostEntityById(request.getId());
+
+        if (!currentUserService.getUsername().equals(post.getAuthor().getUserName())) {
+            throw new AccessDeniedException();
+        }
+
         mapper.update(request, post);
         postRepository.save(post);
     }
@@ -59,6 +66,17 @@ public class PostService {
     @Transactional
     public void delete(UUID postId) {
         Post post = getPostEntityById(postId);
+
+        if (!currentUserService.getUsername().equals(post.getAuthor().getUserName()) && currentUserService.isUserAdminOrHR()) {
+            throw new AccessDeniedException();
+        }
+
+        User currentUser = currentUserService.getCurrentUserEntity();
+        if (!currentUser.getId().equals(post.getAuthor().getId())) {
+            String message = "Your post with title '" + post.getTitle() + "' was deleted by " + currentUser.getUserName();
+            notificationService.createNotification(post.getAuthor().getId(), message);
+        }
+
         postRepository.delete(post);
     }
 
@@ -103,10 +121,9 @@ public class PostService {
     @Transactional
     public void updateComment(UUID postId, UUID commentId, UpdateCommentRequest request) {
         PostComment comment = getCommentEntityById(commentId);
-        User currentUser = currentUserService.getCurrentUserEntity();
 
-        if (!comment.getAuthor().equals(currentUser) && !currentUser.getRoles().stream().anyMatch(role -> "Admin".equals(role.getName()) || "HR".equals(role.getName()))) {
-            throw new UnauthorisedException("You can only update your own comments");
+        if (!currentUserService.getUsername().equals(comment.getAuthor().getUserName())) {
+            throw new AccessDeniedException();
         }
 
         mapper.updateComment(request, comment);
@@ -118,8 +135,13 @@ public class PostService {
         PostComment comment = getCommentEntityById(commentId);
         User currentUser = currentUserService.getCurrentUserEntity();
 
-        if (!comment.getAuthor().equals(currentUser) && !currentUser.getRoles().stream().anyMatch(role -> "Admin".equals(role.getName()) || "HR".equals(role.getName()))) {
-            throw new UnauthorisedException("You can only delete your own comments");
+        if (!currentUserService.getUsername().equals(comment.getAuthor().getUserName()) && currentUserService.isUserAdminOrHR()) {
+            throw new AccessDeniedException();
+        }
+
+        if (!currentUser.getId().equals(comment.getAuthor().getId())) {
+            String message = "Your comment on post '" + comment.getPost().getTitle() + "' was deleted by " + currentUser.getUserName();
+            notificationService.createNotification(comment.getAuthor().getId(), message);
         }
 
         Post post = comment.getPost();
